Fixed Disk II sound problem when pausing/waking up, implemented Apple II cassette out option, implemented Apple II speaker tension/relaxation, restored support for Mac OS X 10.5 (Intel), fixed crash when writing to a DSK file, added PH1 inhibiting writing in Disk II, modified mouse release message to "Press CTRL+OPTION to release mouse."

+++ libemulation-hal/PAAudio.h
@@ -26,7 +26,7 @@ public:
     ~PAAudio();
     
     void setFullDuplex(bool value);
-    void setSampleRate(double value);
+    void setSampleRate(float value);
     void setChannelNum(OEInt value);
     void setFramesPerBuffer(OEInt value);
     void setBufferNum(OEInt value);
@@ -65,7 +65,7 @@ public:
     
 private:
     bool fullDuplex;
-    double sampleRate;
+    float sampleRate;
     OEInt channelNum;
     OEInt framesPerBuffer;
     OEInt bufferNum;
