+++ macosx/PreferencesWindowController.m
@@ -150,9 +150,9 @@
     [view setHidden:YES];
     
     NSRect windowRect = [window frame];
-    float difference = (([view frame].size.height - 
-                         [[window contentView] frame].size.height) *
-                        [window userSpaceScaleFactor]);
+    CGFloat difference = (([view frame].size.height - 
+                           [[window contentView] frame].size.height) *
+                          [window userSpaceScaleFactor]);
     windowRect.origin.y -= difference;
     windowRect.size.height += difference;
